Refactor AI usage backends
