+++ docking/applets/aiusage/applet.py
@@ -2,13 +2,8 @@
 
 from __future__ import annotations
 
-import json
-import os
-import re
-import sys
 from collections.abc import Callable
-from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import gi
 
@@ -18,6 +13,7 @@ gi.require_version("GdkPixbuf", "2.0")
 from gi.repository import Gdk, GdkPixbuf, GLib, Gtk
 
 from docking.applets.aiusage import meta
+from docking.applets.aiusage.backends import BACKENDS
 from docking.applets.aiusage.render import render_icon
 from docking.applets.aiusage.state import (
     DisplayMode,
@@ -31,8 +27,6 @@ from docking.applets.aiusage.state import (
     provider_cost,
     provider_for_model,
     provider_tokens,
-    query_codex_today,
-    query_opencode_today,
     reset_today,
     set_session,
     state_from_prefs,
@@ -41,6 +35,7 @@ from docking.applets.aiusage.state import (
     week_cost,
     week_tokens,
 )
+from docking.applets.aiusage.store import read_prefs_from_disk
 from docking.applets.base import Applet
 from docking.applets.menu import menu_sections, radio_menu_items
 from docking.i18n import _
@@ -55,33 +50,9 @@ REFRESH_INTERVAL_S = 60
 PREFS_KEY = "aiusage"
 PREFS_KEY_LEGACY = "claude"
 
-_CLAUDE_SETTINGS = Path.home() / ".claude" / "settings.json"
-_CODEX_CONFIG = Path.home() / ".codex" / "config.toml"
 
-
-def _project_root() -> str:
-    """Resolve the docking project root from this file's location."""
-    # docking/applets/aiusage/applet.py -> 3 levels up
-    return str(Path(__file__).resolve().parent.parent.parent.parent)
-
-
-def _hook_command_prefix() -> str:
-    root = _project_root()
-    return f"PYTHONPATH={root} {sys.executable} -m docking.applets.aiusage.hook"
-
-
-def _read_prefs_from_disk() -> dict[str, Any] | None:
-    """Read aiusage prefs directly from dock.json on disk."""
-    xdg = os.environ.get("XDG_CONFIG_HOME", "")
-    base = Path(xdg) if xdg else Path.home() / ".config"
-    config_path = base / "docking" / "dock.json"
-    try:
-        config = json.loads(config_path.read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError) as exc:
-        log.debug("Failed to read AI usage prefs from %s: %s", config_path, exc)
-        return None
-    prefs = config.get("applet_prefs", {})
-    return prefs.get(PREFS_KEY) or prefs.get(PREFS_KEY_LEGACY)
+def _read_prefs_from_disk() -> dict | None:
+    return read_prefs_from_disk()
 
 
 class AiUsageApplet(Applet):
@@ -101,8 +72,7 @@ class AiUsageApplet(Applet):
         self._timer_id: int = 0
         self._selected_provider: Provider | None = None
         self._display_mode: DisplayMode = DisplayMode.COST
-        self._codex_poll_error: str | None = None
-        self._opencode_poll_error: str | None = None
+        self._poll_errors: dict[Provider, str] = {}
 
         super().__init__(icon_size, config)
         self.present()
@@ -123,8 +93,8 @@ class AiUsageApplet(Applet):
 
     def start(self, notify: Callable[[], None]) -> None:
         super().start(notify=notify)
-        _register_claude_hooks()
-        _register_codex_hook()
+        for backend in BACKENDS:
+            backend.register_hooks()
         self._tick()  # Immediate first poll (merges OpenCode etc.).
         self._timer_id = GLib.timeout_add_seconds(REFRESH_INTERVAL_S, self._tick)
 
@@ -138,7 +108,7 @@ class AiUsageApplet(Applet):
         pass
 
     def on_scroll(self, direction_up: bool) -> None:
-        """Cycle through providers: Auto -> Claude -> Codex."""
+        """Cycle through providers: Auto -> Claude -> Codex -> OpenCode."""
         choices: list[Provider | None] = [
             None,
             Provider.CLAUDE,
@@ -197,40 +167,25 @@ class AiUsageApplet(Applet):
         prefs = _read_prefs_from_disk()
         new_state = state_from_prefs(prefs=prefs)
 
-        try:
-            codex_sessions = query_codex_today()
-            self._codex_poll_error = None
-            for sid, model_usage in codex_sessions.items():
-                new_state = set_session(
-                    state=new_state,
-                    session_id=sid,
-                    model_usage=model_usage,
-                )
-        except Exception as exc:
-            error = f"{type(exc).__name__}: {exc}"
-            if error != self._codex_poll_error:
-                self._codex_poll_error = error
-                log.bind(action="poll_codex").warning(
-                    "Failed to poll Codex usage: %s", error
-                )
-
-        # Merge OpenCode sessions from SQLite (no hook, poll-based).
-        try:
-            oc_sessions = query_opencode_today()
-            self._opencode_poll_error = None
-            for sid, model_usage in oc_sessions.items():
-                new_state = set_session(
-                    state=new_state,
-                    session_id=f"oc:{sid}",
-                    model_usage=model_usage,
-                )
-        except Exception as exc:
-            error = f"{type(exc).__name__}: {exc}"
-            if error != self._opencode_poll_error:
-                self._opencode_poll_error = error
-                log.bind(action="poll_opencode").warning(
-                    "Failed to poll OpenCode usage: %s", error
-                )
+        for backend in BACKENDS:
+            try:
+                sessions = backend.poll_today()
+                self._poll_errors.pop(backend.provider, None)
+                for sid, model_usage in sessions.items():
+                    new_state = set_session(
+                        state=new_state,
+                        session_id=sid,
+                        model_usage=model_usage,
+                    )
+            except Exception as exc:
+                error = f"{type(exc).__name__}: {exc}"
+                if error != self._poll_errors.get(backend.provider):
+                    self._poll_errors[backend.provider] = error
+                    log.bind(action=f"poll_{backend.provider.value}").warning(
+                        "Failed to poll %s usage: %s",
+                        backend.provider.value,
+                        error,
+                    )
 
         if new_state != self._state:
             self._state = new_state
@@ -254,12 +209,12 @@ class AiUsageApplet(Applet):
             box.pack_start(label, False, False, 0)
             return box
 
-        # Filter models to selected provider and skip zero-usage.
+        # Filter models to selected provider and skip empty rows.
         models = [
             (m, u)
             for m, u in entry.by_model
             if (sel is None or provider_for_model(model=m) == sel)
-            and cost_for_usage(model=m, usage=u) > 0
+            and (cost_for_usage(model=m, usage=u) > 0 or total_tokens(u) > 0)
         ]
 
         if not models:
@@ -331,139 +286,3 @@ class AiUsageApplet(Applet):
             box.pack_start(week_lbl, False, False, 0)
 
         return box
-
-
-# ------------------------------------------------------------------
-# Claude hook registration
-# ------------------------------------------------------------------
-
-
-def _register_claude_hooks() -> None:
-    """Ensure Claude Code hooks point to our CLI entry point."""
-    try:
-        if _CLAUDE_SETTINGS.exists():
-            settings = json.loads(_CLAUDE_SETTINGS.read_text(encoding="utf-8"))
-        else:
-            settings = {}
-    except (OSError, json.JSONDecodeError):
-        log.bind(action="register_hooks").warning("Could not read %s", _CLAUDE_SETTINGS)
-        return
-
-    hooks = settings.setdefault("hooks", {})
-    changed = False
-    prefix = _hook_command_prefix()
-
-    # Remove stale claude.hook entries.
-    for event_key in ("Stop", "SessionStart"):
-        entries = hooks.get(event_key, [])
-        cleaned = [
-            e
-            for e in entries
-            if not any(
-                "docking.applets.claude.hook" in h.get("command", "")
-                for h in e.get("hooks", [])
-            )
-        ]
-        if len(cleaned) != len(entries):
-            hooks[event_key] = cleaned
-            changed = True
-
-    # Stop hook (needs matcher).
-    stop_entries = hooks.get("Stop", [])
-    if not _has_hook(entries=stop_entries, needle=prefix):
-        stop_entries.append(
-            {
-                "matcher": "*",
-                "hooks": [{"type": "command", "command": f"{prefix} claude Stop"}],
-            }
-        )
-        hooks["Stop"] = stop_entries
-        changed = True
-
-    # SessionStart hook.
-    start_entries = hooks.get("SessionStart", [])
-    if not _has_hook(entries=start_entries, needle=prefix):
-        start_entries.append(
-            {
-                "hooks": [
-                    {"type": "command", "command": f"{prefix} claude SessionStart"}
-                ],
-            }
-        )
-        hooks["SessionStart"] = start_entries
-        changed = True
-
-    if changed:
-        try:
-            _CLAUDE_SETTINGS.parent.mkdir(parents=True, exist_ok=True)
-            _CLAUDE_SETTINGS.write_text(
-                json.dumps(settings, indent=2) + "\n", encoding="utf-8"
-            )
-        except OSError:
-            log.bind(action="register_hooks").warning(
-                "Could not write %s", _CLAUDE_SETTINGS
-            )
-
-
-def _has_hook(entries: list[dict], needle: str) -> bool:
-    for entry in entries:
-        for h in entry.get("hooks", []):
-            if needle in h.get("command", ""):
-                return True
-    return False
-
-
-# ------------------------------------------------------------------
-# Codex hook registration
-# ------------------------------------------------------------------
-
-
-def _register_codex_hook() -> None:
-    """Ensure Codex CLI notify points to our hook."""
-    try:
-        if _CODEX_CONFIG.exists():
-            content = _CODEX_CONFIG.read_text(encoding="utf-8")
-        else:
-            return  # Codex not installed.
-    except OSError as exc:
-        log.debug("Failed to read Codex config %s: %s", _CODEX_CONFIG, exc)
-        return
-
-    root = _project_root()
-    our_toml = (
-        f'notify = ["env", "PYTHONPATH={root}",'
-        f' "{sys.executable}", "-m", "docking.applets.aiusage.hook", "codex"]'
-    )
-
-    notify_match = re.search(r"^notify\s*=\s*\[.*?\]", content, re.MULTILINE)
-    if notify_match:
-        existing = notify_match.group(0)
-        if "PYTHONPATH" in existing and "docking.applets.aiusage.hook" in existing:
-            return  # Already ours with correct PYTHONPATH.
-        if "codex-sync" in existing:
-            log.bind(action="register_codex_hook").info(
-                "Codex notify already set to codex-sync, not overwriting"
-            )
-            return
-        content = (
-            content[: notify_match.start()] + our_toml + content[notify_match.end() :]
-        )
-    else:
-        # Insert before first [section].
-        section_match = re.search(r"^\[", content, re.MULTILINE)
-        if section_match:
-            content = (
-                content[: section_match.start()]
-                + our_toml
-                + "\n"
-                + content[section_match.start() :]
-            )
-        else:
-            content = content.rstrip() + "\n" + our_toml + "\n"
-
-    try:
-        _CODEX_CONFIG.write_text(content, encoding="utf-8")
-    except OSError:
-        log.bind(action="register_codex_hook").warning(
-            "Could not write %s", _CODEX_CONFIG
-        )
