+++ docking/applets/aiusage/state.py
@@ -3,19 +3,15 @@
 from __future__ import annotations
 
 import datetime
-import json
 from collections.abc import Mapping
 from dataclasses import dataclass, replace
 from enum import Enum
-from pathlib import Path
 from typing import Any
 
 from docking.applets.tooltip import structured_tooltip
 from docking.i18n import _
-from docking.log import get_logger
 
 MAX_DAYS = 7
-log = get_logger("aiusage.state")
 
 
 class Provider(str, Enum):
@@ -481,238 +477,3 @@ def _has_usage(u: ModelUsage) -> bool:
         or u.cache_read_tokens
         or u.precalculated_cost
     )
-
-
-def parse_claude_transcript(path: Path) -> dict[str, ModelUsage]:
-    """Read a Claude Code JSONL transcript and accumulate per-model usage."""
-    result: dict[str, ModelUsage] = {}
-    try:
-        text = path.read_text(encoding="utf-8")
-    except OSError as exc:
-        log.debug("Failed to read Claude transcript %s: %s", path, exc)
-        return result
-
-    for line in text.splitlines():
-        try:
-            entry = json.loads(line)
-        except (json.JSONDecodeError, ValueError) as exc:
-            log.debug("Failed to parse Claude transcript line in %s: %s", path, exc)
-            continue
-        if entry.get("type") != "assistant":
-            continue
-        msg = entry.get("message")
-        if not isinstance(msg, dict):
-            continue
-        usage = msg.get("usage")
-        if not isinstance(usage, dict):
-            continue
-        model = msg.get("model", "")
-        if not model:
-            continue
-
-        prev = result.get(model, ModelUsage())
-        result[model] = ModelUsage(
-            input_tokens=prev.input_tokens + int(usage.get("input_tokens", 0)),
-            output_tokens=prev.output_tokens + int(usage.get("output_tokens", 0)),
-            cache_write_tokens=prev.cache_write_tokens
-            + int(usage.get("cache_creation_input_tokens", 0)),
-            cache_read_tokens=prev.cache_read_tokens
-            + int(usage.get("cache_read_input_tokens", 0)),
-        )
-    return {m: u for m, u in result.items() if _has_usage(u)}
-
-
-# ---------------------------------------------------------------------------
-# Codex transcript parsing
-# ---------------------------------------------------------------------------
-
-
-def find_codex_session(thread_id: str | None = None) -> Path | None:
-    """Find a Codex session JSONL file by thread-id or most recent."""
-    sessions_dir = Path.home() / ".codex" / "sessions"
-    if not sessions_dir.is_dir():
-        return None
-
-    if thread_id:
-        for jsonl in sorted(sessions_dir.rglob("*.jsonl"), reverse=True):
-            if thread_id in jsonl.name:
-                return jsonl
-
-    # Fallback: most recent by mtime.
-    candidates = sorted(
-        sessions_dir.rglob("*.jsonl"),
-        key=lambda p: p.stat().st_mtime,
-        reverse=True,
-    )
-    return candidates[0] if candidates else None
-
-
-def parse_codex_transcript(path: Path) -> dict[str, ModelUsage]:
-    """Read a Codex CLI JSONL session and extract cumulative usage."""
-    try:
-        text = path.read_text(encoding="utf-8")
-    except OSError as exc:
-        log.debug("Failed to read Codex transcript %s: %s", path, exc)
-        return {}
-
-    model: str = ""
-    best_total: int = 0
-    best_usage: dict[str, int] = {}
-
-    for line in text.splitlines():
-        try:
-            entry = json.loads(line)
-        except (json.JSONDecodeError, ValueError) as exc:
-            log.debug("Failed to parse Codex transcript line in %s: %s", path, exc)
-            continue
-
-        entry_type = entry.get("type", "")
-
-        if entry_type == "turn_context":
-            m = entry.get("payload", {}).get("model", "")
-            if m:
-                model = m
-
-        if entry_type == "event_msg":
-            payload = entry.get("payload", {})
-            if payload.get("type") != "token_count":
-                continue
-            info = payload.get("info")
-            if not isinstance(info, dict):
-                continue
-            total_usage = info.get("total_token_usage", {})
-            total = int(total_usage.get("total_tokens", 0))
-            if total > best_total:
-                best_total = total
-                best_usage = total_usage
-
-    if not model or not best_usage:
-        return {}
-
-    return {
-        model: ModelUsage(
-            input_tokens=int(best_usage.get("input_tokens", 0)),
-            output_tokens=int(best_usage.get("output_tokens", 0)),
-            cache_write_tokens=0,
-            cache_read_tokens=int(best_usage.get("cached_input_tokens", 0)),
-        )
-    }
-
-
-def _codex_session_id(path: Path) -> str:
-    try:
-        for line in path.read_text(encoding="utf-8").splitlines():
-            try:
-                entry = json.loads(line)
-            except (json.JSONDecodeError, ValueError):
-                continue
-            if entry.get("type") != "session_meta":
-                continue
-            session_id = entry.get("payload", {}).get("id", "")
-            if session_id:
-                return str(session_id)
-    except OSError as exc:
-        log.debug("Failed to read Codex session metadata %s: %s", path, exc)
-    return path.stem
-
-
-def query_codex_today() -> dict[str, dict[str, ModelUsage]]:
-    """Read today's Codex sessions from local JSONL transcripts.
-
-    Codex ``notify`` hooks only apply to processes that started after the hook
-    was registered. Polling recent session files lets the applet recover usage
-    for already-running Codex sessions as well.
-    """
-    sessions_dir = Path.home() / ".codex" / "sessions"
-    if not sessions_dir.is_dir():
-        return {}
-
-    today_start = datetime.datetime.fromisoformat(_today_iso()).timestamp()
-    result: dict[str, dict[str, ModelUsage]] = {}
-    for jsonl in sessions_dir.rglob("*.jsonl"):
-        try:
-            if jsonl.stat().st_mtime < today_start:
-                continue
-        except OSError as exc:
-            log.debug("Failed to stat Codex transcript %s: %s", jsonl, exc)
-            continue
-
-        model_usage = parse_codex_transcript(path=jsonl)
-        if model_usage:
-            result[_codex_session_id(path=jsonl)] = model_usage
-    return result
-
-
-# ---------------------------------------------------------------------------
-# OpenCode usage (SQLite database)
-# ---------------------------------------------------------------------------
-
-_OPENCODE_DB = Path.home() / ".local" / "share" / "opencode" / "opencode.db"
-
-
-def query_opencode_today() -> dict[str, dict[str, ModelUsage]]:
-    """Query OpenCode SQLite for today's sessions.
-
-    Returns {session_id: {prefixed_model: ModelUsage}}.
-    """
-    import sqlite3
-
-    if not _OPENCODE_DB.exists():
-        return {}
-
-    today = _today_iso()
-    # time_created is Unix ms; compute start-of-day in ms.
-    today_start_ms = int(datetime.datetime.fromisoformat(today).timestamp() * 1000)
-
-    try:
-        conn = sqlite3.connect(f"file:{_OPENCODE_DB}?mode=ro", uri=True)
-    except sqlite3.OperationalError as exc:
-        log.debug("Failed to open OpenCode database %s: %s", _OPENCODE_DB, exc)
-        return {}
-
-    try:
-        cur = conn.cursor()
-        cur.execute(
-            "SELECT id FROM session WHERE time_created >= ?",
-            (today_start_ms,),
-        )
-        session_ids = [r[0] for r in cur.fetchall()]
-        if not session_ids:
-            return {}
-
-        result: dict[str, dict[str, ModelUsage]] = {}
-        for sid in session_ids:
-            cur.execute("SELECT data FROM message WHERE session_id = ?", (sid,))
-            by_model: dict[str, ModelUsage] = {}
-            for (data_str,) in cur.fetchall():
-                try:
-                    data = json.loads(data_str)
-                except (json.JSONDecodeError, ValueError) as exc:
-                    log.debug(
-                        "Failed to parse OpenCode message row for session %s: %s",
-                        sid,
-                        exc,
-                    )
-                    continue
-                model_id = data.get("modelID", "")
-                if not model_id:
-                    continue
-                tokens = data.get("tokens", {})
-                cost = float(data.get("cost", 0) or 0)
-                inp = int(tokens.get("input", 0))
-                out = int(tokens.get("output", 0))
-                if not (inp or out or cost):
-                    continue
-
-                key = f"{OPENCODE_PREFIX}{model_id}"
-                prev = by_model.get(key, ModelUsage())
-                by_model[key] = ModelUsage(
-                    input_tokens=prev.input_tokens + inp,
-                    output_tokens=prev.output_tokens + out,
-                    precalculated_cost=prev.precalculated_cost + cost,
-                )
-            if by_model:
-                result[sid] = by_model
-        return result
-    finally:
-        conn.close()
