+++ tests/applets/test_aiusage.py
@@ -13,8 +13,11 @@ from types import SimpleNamespace
 from unittest.mock import MagicMock
 
 import docking.applets.aiusage.applet as aiusage_mod
+import docking.applets.aiusage.backends.claude as claude_backend
+import docking.applets.aiusage.backends.codex as codex_backend
+import docking.applets.aiusage.backends.opencode as opencode_backend
 import docking.applets.aiusage.render as aiusage_render_mod
-import docking.applets.aiusage.state as aiusage_state_mod
+import docking.applets.aiusage.store as aiusage_store
 from docking.applets.aiusage.applet import AiUsageApplet
 from docking.applets.aiusage.state import (
     AiUsageState,
@@ -26,16 +29,11 @@ from docking.applets.aiusage.state import (
     day_cost,
     day_tokens,
     dominant_provider,
-    find_codex_session,
     match_model_tier,
-    parse_claude_transcript,
-    parse_codex_transcript,
     prefs_from_state,
     provider_cost,
     provider_for_model,
     provider_tokens,
-    query_codex_today,
-    query_opencode_today,
     reset_today,
     set_session,
     state_from_prefs,
@@ -446,16 +444,16 @@ class TestClaudeTranscript:
             ),
         ]
         jsonl.write_text("\n".join(lines))
-        result = parse_claude_transcript(path=jsonl)
+        result = claude_backend.parse_transcript(path=jsonl)
         assert result["claude-opus-4-20250514"].input_tokens == 400
 
     def test_missing_file(self, tmp_path):
-        assert parse_claude_transcript(path=tmp_path / "x.jsonl") == {}
+        assert claude_backend.parse_transcript(path=tmp_path / "x.jsonl") == {}
 
     def test_ignores_non_assistant(self, tmp_path):
         jsonl = tmp_path / "session.jsonl"
         jsonl.write_text(json.dumps({"type": "user", "message": {"role": "user"}}))
-        assert parse_claude_transcript(path=jsonl) == {}
+        assert claude_backend.parse_transcript(path=jsonl) == {}
 
     def test_ignores_malformed_and_zero_usage_entries(self, tmp_path):
         jsonl = tmp_path / "session.jsonl"
@@ -481,7 +479,7 @@ class TestClaudeTranscript:
             )
         )
 
-        assert parse_claude_transcript(path=jsonl) == {}
+        assert claude_backend.parse_transcript(path=jsonl) == {}
 
 
 # ---------------------------------------------------------------
@@ -493,7 +491,7 @@ class TestCodexTranscript:
     def test_find_codex_session_missing_dir(self, tmp_path, monkeypatch):
         monkeypatch.setattr(Path, "home", lambda: tmp_path)
 
-        assert find_codex_session() is None
+        assert codex_backend.find_session() is None
 
     def test_find_codex_session_by_thread_or_mtime(self, tmp_path, monkeypatch):
         sessions_dir = tmp_path / ".codex" / "sessions"
@@ -506,8 +504,8 @@ class TestCodexTranscript:
         os.utime(newer, (2000, 2000))
         monkeypatch.setattr(Path, "home", lambda: tmp_path)
 
-        assert find_codex_session(thread_id="thread-123") == newer
-        assert find_codex_session() == newer
+        assert codex_backend.find_session(thread_id="thread-123") == newer
+        assert codex_backend.find_session() == newer
 
     def test_parses_valid_session(self, tmp_path):
         jsonl = tmp_path / "session.jsonl"
@@ -536,7 +534,7 @@ class TestCodexTranscript:
             ),
         ]
         jsonl.write_text("\n".join(lines))
-        result = parse_codex_transcript(path=jsonl)
+        result = codex_backend.parse_transcript(path=jsonl)
         assert "gpt-5.4" in result
         u = result["gpt-5.4"]
         assert u.input_tokens == 1000
@@ -581,11 +579,11 @@ class TestCodexTranscript:
             ),
         ]
         jsonl.write_text("\n".join(lines))
-        result = parse_codex_transcript(path=jsonl)
+        result = codex_backend.parse_transcript(path=jsonl)
         assert result["gpt-5.4"].input_tokens == 500
 
     def test_missing_file(self, tmp_path):
-        assert parse_codex_transcript(path=tmp_path / "x.jsonl") == {}
+        assert codex_backend.parse_transcript(path=tmp_path / "x.jsonl") == {}
 
     def test_no_model_returns_empty(self, tmp_path):
         jsonl = tmp_path / "session.jsonl"
@@ -600,7 +598,7 @@ class TestCodexTranscript:
                 }
             )
         )
-        assert parse_codex_transcript(path=jsonl) == {}
+        assert codex_backend.parse_transcript(path=jsonl) == {}
 
     def test_ignores_malformed_and_non_token_entries(self, tmp_path):
         jsonl = tmp_path / "session.jsonl"
@@ -620,14 +618,14 @@ class TestCodexTranscript:
             )
         )
 
-        assert parse_codex_transcript(path=jsonl) == {}
+        assert codex_backend.parse_transcript(path=jsonl) == {}
 
     def test_codex_session_id_falls_back_to_stem(self, tmp_path):
         jsonl = tmp_path / "session.jsonl"
         jsonl.write_text("{\n{}\n", encoding="utf-8")
 
-        assert aiusage_state_mod._codex_session_id(jsonl) == "session"
-        assert aiusage_state_mod._codex_session_id(tmp_path / "missing") == "missing"
+        assert codex_backend.session_id(jsonl) == "session"
+        assert codex_backend.session_id(tmp_path / "missing") == "missing"
 
     def test_query_codex_today_reads_recent_sessions(self, tmp_path, monkeypatch):
         sessions_dir = tmp_path / ".codex" / "sessions" / "2026" / "04" / "29"
@@ -671,11 +669,11 @@ class TestCodexTranscript:
         os.utime(jsonl, (today_ts, today_ts))
         monkeypatch.setattr(Path, "home", lambda: tmp_path)
         monkeypatch.setattr(
-            "docking.applets.aiusage.state._today_iso",
+            "docking.applets.aiusage.backends.codex._today_iso",
             lambda: "2026-04-29",
         )
 
-        result = query_codex_today()
+        result = codex_backend.query_today()
 
         assert result["thread-1"]["gpt-5.5"].input_tokens == 1200
 
@@ -692,30 +690,30 @@ class TestCodexTranscript:
         os.utime(current, (2000, 2000))
         monkeypatch.setattr(Path, "home", lambda: tmp_path)
         monkeypatch.setattr(
-            "docking.applets.aiusage.state._today_iso",
+            "docking.applets.aiusage.backends.codex._today_iso",
             lambda: "1970-01-01",
         )
 
-        assert query_codex_today() == {}
+        assert codex_backend.query_today() == {}
 
 
 class TestOpenCodeUsage:
     def test_query_opencode_today_missing_db(self, tmp_path, monkeypatch):
-        monkeypatch.setattr(aiusage_state_mod, "_OPENCODE_DB", tmp_path / "missing.db")
+        monkeypatch.setattr(opencode_backend, "_OPENCODE_DB", tmp_path / "missing.db")
 
-        assert query_opencode_today() == {}
+        assert opencode_backend.query_today() == {}
 
     def test_query_opencode_today_connect_error(self, tmp_path, monkeypatch):
         db = tmp_path / "opencode.db"
         db.write_text("", encoding="utf-8")
-        monkeypatch.setattr(aiusage_state_mod, "_OPENCODE_DB", db)
+        monkeypatch.setattr(opencode_backend, "_OPENCODE_DB", db)
 
         def fail(*_args, **_kwargs):
             raise sqlite3.OperationalError("locked")
 
         monkeypatch.setattr(sqlite3, "connect", fail)
 
-        assert query_opencode_today() == {}
+        assert opencode_backend.query_today() == {}
 
     def test_query_opencode_today_no_sessions(self, tmp_path, monkeypatch):
         db = tmp_path / "opencode.db"
@@ -724,10 +722,10 @@ class TestOpenCodeUsage:
         conn.execute("CREATE TABLE message (session_id TEXT, data TEXT)")
         conn.commit()
         conn.close()
-        monkeypatch.setattr(aiusage_state_mod, "_OPENCODE_DB", db)
-        monkeypatch.setattr(aiusage_state_mod, "_today_iso", lambda: "2026-04-29")
+        monkeypatch.setattr(opencode_backend, "_OPENCODE_DB", db)
+        monkeypatch.setattr(opencode_backend, "_today_iso", lambda: "2026-04-29")
 
-        assert query_opencode_today() == {}
+        assert opencode_backend.query_today() == {}
 
     def test_query_opencode_today_accumulates_valid_rows(self, tmp_path, monkeypatch):
         db = tmp_path / "opencode.db"
@@ -763,12 +761,12 @@ class TestOpenCodeUsage:
         conn.executemany("INSERT INTO message VALUES (?, ?)", [("s1", r) for r in rows])
         conn.commit()
         conn.close()
-        monkeypatch.setattr(aiusage_state_mod, "_OPENCODE_DB", db)
-        monkeypatch.setattr(aiusage_state_mod, "_today_iso", lambda: "2026-04-29")
+        monkeypatch.setattr(opencode_backend, "_OPENCODE_DB", db)
+        monkeypatch.setattr(opencode_backend, "_today_iso", lambda: "2026-04-29")
 
-        result = query_opencode_today()
+        result = opencode_backend.query_today()
 
-        usage = result["s1"]["opencode:m"]
+        usage = result["oc:s1"]["opencode:m"]
         assert usage.input_tokens == 12
         assert usage.output_tokens == 8
         assert usage.precalculated_cost == 0.35
@@ -797,10 +795,11 @@ class TestHook:
         config_path = tmp_path / "dock.json"
         config_path.write_text("{}")
 
-        from docking.applets.aiusage import hook
-
-        monkeypatch.setattr(hook, "_config_path", lambda: config_path)
-        hook._handle_claude_stop(data={"transcript_path": str(jsonl)})
+        monkeypatch.setattr(aiusage_store, "config_path", lambda: config_path)
+        claude_backend.ClaudeBackend().handle_hook(
+            event="Stop",
+            payload={"transcript_path": str(jsonl)},
+        )
 
         config = json.loads(config_path.read_text())
         prefs = config["applet_prefs"]["aiusage"]
@@ -833,13 +832,10 @@ class TestHook:
         config_path = tmp_path / "dock.json"
         config_path.write_text("{}")
 
-        from docking.applets.aiusage import hook
-        from docking.applets.aiusage import state as state_mod
+        monkeypatch.setattr(aiusage_store, "config_path", lambda: config_path)
+        monkeypatch.setattr(codex_backend, "find_session", lambda thread_id: jsonl)
 
-        monkeypatch.setattr(hook, "_config_path", lambda: config_path)
-        monkeypatch.setattr(state_mod, "find_codex_session", lambda thread_id: jsonl)
-
-        hook._handle_codex_turn(json_arg=json.dumps({"thread-id": "test"}))
+        codex_backend.handle_turn(json_arg=json.dumps({"thread-id": "test"}))
 
         config = json.loads(config_path.read_text())
         prefs = config["applet_prefs"]["aiusage"]
@@ -849,10 +845,8 @@ class TestHook:
         config_path = tmp_path / "dock.json"
         config_path.write_text("{}")
 
-        from docking.applets.aiusage import hook
-
-        monkeypatch.setattr(hook, "_config_path", lambda: config_path)
-        hook._handle_claude_stop(data={})
+        monkeypatch.setattr(claude_backend, "replace_session", MagicMock())
+        claude_backend.ClaudeBackend().handle_hook(event="Stop", payload={})
         assert json.loads(config_path.read_text()) == {}
 
 
@@ -915,8 +909,17 @@ class TestAiUsageApplet:
             "source_remove",
             lambda timer_id: removed.append(timer_id),
         )
-        monkeypatch.setattr(aiusage_mod, "_register_claude_hooks", lambda: None)
-        monkeypatch.setattr(aiusage_mod, "_register_codex_hook", lambda: None)
+        monkeypatch.setattr(
+            aiusage_mod,
+            "BACKENDS",
+            (
+                SimpleNamespace(
+                    provider=Provider.CLAUDE,
+                    register_hooks=lambda: None,
+                    poll_today=dict,
+                ),
+            ),
+        )
 
         applet.start(lambda: None)
         assert applet._timer_id == 999
@@ -1010,23 +1013,28 @@ class TestAiUsageApplet:
         applet = AiUsageApplet(48)
         applet.present = MagicMock()
         monkeypatch.setattr(aiusage_mod, "_read_prefs_from_disk", lambda: None)
-        monkeypatch.setattr(aiusage_mod, "query_codex_today", dict)
         monkeypatch.setattr(
             aiusage_mod,
-            "query_opencode_today",
-            lambda: {
-                "abc": {
-                    "opencode:gpt-oss": ModelUsage(
-                        input_tokens=10,
-                        output_tokens=5,
-                        precalculated_cost=1.25,
-                    )
-                }
-            },
+            "BACKENDS",
+            (
+                SimpleNamespace(
+                    provider=Provider.OPENCODE,
+                    register_hooks=lambda: None,
+                    poll_today=lambda: {
+                        "abc": {
+                            "opencode:gpt-oss": ModelUsage(
+                                input_tokens=10,
+                                output_tokens=5,
+                                precalculated_cost=1.25,
+                            )
+                        }
+                    },
+                ),
+            ),
         )
 
         assert applet._tick() is True
-        assert applet._opencode_poll_error is None
+        assert applet._poll_errors == {}
         assert applet._state.days[0].sessions == 1
         assert applet.present.call_count == 1
 
@@ -1034,23 +1042,28 @@ class TestAiUsageApplet:
         applet = AiUsageApplet(48)
         applet.present = MagicMock()
         monkeypatch.setattr(aiusage_mod, "_read_prefs_from_disk", lambda: None)
-        monkeypatch.setattr(aiusage_mod, "query_opencode_today", dict)
         monkeypatch.setattr(
             aiusage_mod,
-            "query_codex_today",
-            lambda: {
-                "thread-1": {
-                    "gpt-5.5": ModelUsage(
-                        input_tokens=1200,
-                        output_tokens=50,
-                        cache_read_tokens=200,
-                    )
-                }
-            },
+            "BACKENDS",
+            (
+                SimpleNamespace(
+                    provider=Provider.CODEX,
+                    register_hooks=lambda: None,
+                    poll_today=lambda: {
+                        "thread-1": {
+                            "gpt-5.5": ModelUsage(
+                                input_tokens=1200,
+                                output_tokens=50,
+                                cache_read_tokens=200,
+                            )
+                        }
+                    },
+                ),
+            ),
         )
 
         assert applet._tick() is True
-        assert applet._codex_poll_error is None
+        assert applet._poll_errors == {}
         assert applet._state.days[0].sessions == 1
         assert applet._state.days[0].by_model[0][0] == "gpt-5.5"
         assert applet.present.call_count == 1
@@ -1074,18 +1087,27 @@ class TestAiUsageApplet:
     def test_tick_warns_once_when_opencode_poll_fails(self, monkeypatch, caplog):
         applet = AiUsageApplet(48)
         monkeypatch.setattr(aiusage_mod, "_read_prefs_from_disk", lambda: None)
-        monkeypatch.setattr(aiusage_mod, "query_codex_today", dict)
 
         def fail_query():
             raise RuntimeError("database locked")
 
-        monkeypatch.setattr(aiusage_mod, "query_opencode_today", fail_query)
+        monkeypatch.setattr(
+            aiusage_mod,
+            "BACKENDS",
+            (
+                SimpleNamespace(
+                    provider=Provider.OPENCODE,
+                    register_hooks=lambda: None,
+                    poll_today=fail_query,
+                ),
+            ),
+        )
 
         with caplog.at_level(logging.WARNING, logger="docking.aiusage"):
             assert applet._tick() is True
             assert applet._tick() is True
 
-        assert caplog.text.count("Failed to poll OpenCode usage") == 1
+        assert caplog.text.count("Failed to poll opencode usage") == 1
 
 
 # ---------------------------------------------------------------
@@ -1109,9 +1131,9 @@ class TestClaudeHookRegistration:
         settings_path = tmp_path / ".claude" / "settings.json"
         settings_path.parent.mkdir(parents=True)
         settings_path.write_text("{}")
-        monkeypatch.setattr(aiusage_mod, "_CLAUDE_SETTINGS", settings_path)
+        monkeypatch.setattr(claude_backend, "_CLAUDE_SETTINGS", settings_path)
 
-        aiusage_mod._register_claude_hooks()
+        claude_backend.register_hooks()
 
         settings = json.loads(settings_path.read_text())
         assert "Stop" in settings["hooks"]
@@ -1131,9 +1153,9 @@ class TestClaudeHookRegistration:
             }
         }
         settings_path.write_text(json.dumps(existing))
-        monkeypatch.setattr(aiusage_mod, "_CLAUDE_SETTINGS", settings_path)
+        monkeypatch.setattr(claude_backend, "_CLAUDE_SETTINGS", settings_path)
 
-        aiusage_mod._register_claude_hooks()
+        claude_backend.register_hooks()
 
         settings = json.loads(settings_path.read_text())
         assert len(settings["hooks"]["Stop"]) == 2
@@ -1142,10 +1164,10 @@ class TestClaudeHookRegistration:
         settings_path = tmp_path / ".claude" / "settings.json"
         settings_path.parent.mkdir(parents=True)
         settings_path.write_text("{}")
-        monkeypatch.setattr(aiusage_mod, "_CLAUDE_SETTINGS", settings_path)
+        monkeypatch.setattr(claude_backend, "_CLAUDE_SETTINGS", settings_path)
 
-        aiusage_mod._register_claude_hooks()
-        aiusage_mod._register_claude_hooks()
+        claude_backend.register_hooks()
+        claude_backend.register_hooks()
 
         settings = json.loads(settings_path.read_text())
         assert len(settings["hooks"]["Stop"]) == 1
@@ -1155,10 +1177,10 @@ class TestClaudeHookRegistration:
         settings_path.parent.mkdir(parents=True)
         settings_path.write_text("{")
         logger = SimpleNamespace(warning=MagicMock())
-        monkeypatch.setattr(aiusage_mod, "_CLAUDE_SETTINGS", settings_path)
-        monkeypatch.setattr(aiusage_mod.log, "bind", lambda **_kwargs: logger)
+        monkeypatch.setattr(claude_backend, "_CLAUDE_SETTINGS", settings_path)
+        monkeypatch.setattr(claude_backend.log, "bind", lambda **_kwargs: logger)
 
-        aiusage_mod._register_claude_hooks()
+        claude_backend.register_hooks()
 
         logger.warning.assert_called_once()
 
@@ -1167,8 +1189,8 @@ class TestClaudeHookRegistration:
         settings_path.parent.mkdir(parents=True)
         settings_path.write_text("{}")
         logger = SimpleNamespace(warning=MagicMock())
-        monkeypatch.setattr(aiusage_mod, "_CLAUDE_SETTINGS", settings_path)
-        monkeypatch.setattr(aiusage_mod.log, "bind", lambda **_kwargs: logger)
+        monkeypatch.setattr(claude_backend, "_CLAUDE_SETTINGS", settings_path)
+        monkeypatch.setattr(claude_backend.log, "bind", lambda **_kwargs: logger)
         original_write_text = Path.write_text
 
         def fail_write_text(self, *args, **kwargs):
@@ -1178,12 +1200,12 @@ class TestClaudeHookRegistration:
 
         monkeypatch.setattr(Path, "write_text", fail_write_text)
 
-        aiusage_mod._register_claude_hooks()
+        claude_backend.register_hooks()
 
         logger.warning.assert_called_once()
 
     def test_has_hook_detects_existing_command(self):
-        assert aiusage_mod._has_hook(
+        assert claude_backend._has_hook(
             entries=[{"hooks": [{"command": "prefix value"}]}],
             needle="prefix",
         )
@@ -1193,9 +1215,9 @@ class TestCodexHookRegistration:
     def test_registers_notify(self, tmp_path, monkeypatch):
         config_path = tmp_path / "config.toml"
         config_path.write_text('model = "gpt-5.4"\n')
-        monkeypatch.setattr(aiusage_mod, "_CODEX_CONFIG", config_path)
+        monkeypatch.setattr(codex_backend, "_CODEX_CONFIG", config_path)
 
-        aiusage_mod._register_codex_hook()
+        codex_backend.register_hook()
 
         content = config_path.read_text()
         assert "docking.applets.aiusage.hook" in content
@@ -1206,9 +1228,9 @@ class TestCodexHookRegistration:
         config_path.write_text(
             'model = "gpt-5"\nnotify = ["codex-sync", "hook", "agent-turn-complete"]\n'
         )
-        monkeypatch.setattr(aiusage_mod, "_CODEX_CONFIG", config_path)
+        monkeypatch.setattr(codex_backend, "_CODEX_CONFIG", config_path)
 
-        aiusage_mod._register_codex_hook()
+        codex_backend.register_hook()
 
         content = config_path.read_text()
         assert "codex-sync" in content
@@ -1217,10 +1239,10 @@ class TestCodexHookRegistration:
     def test_idempotent(self, tmp_path, monkeypatch):
         config_path = tmp_path / "config.toml"
         config_path.write_text('model = "gpt-5.4"\n')
-        monkeypatch.setattr(aiusage_mod, "_CODEX_CONFIG", config_path)
+        monkeypatch.setattr(codex_backend, "_CODEX_CONFIG", config_path)
 
-        aiusage_mod._register_codex_hook()
-        aiusage_mod._register_codex_hook()
+        codex_backend.register_hook()
+        codex_backend.register_hook()
 
         content = config_path.read_text()
         assert content.count("notify") == 1
@@ -1228,9 +1250,9 @@ class TestCodexHookRegistration:
     def test_inserts_notify_before_first_section(self, tmp_path, monkeypatch):
         config_path = tmp_path / "config.toml"
         config_path.write_text('[profiles]\ndefault = "x"\n')
-        monkeypatch.setattr(aiusage_mod, "_CODEX_CONFIG", config_path)
+        monkeypatch.setattr(codex_backend, "_CODEX_CONFIG", config_path)
 
-        aiusage_mod._register_codex_hook()
+        codex_backend.register_hook()
 
         content = config_path.read_text()
         assert content.startswith("notify = ")
@@ -1240,8 +1262,8 @@ class TestCodexHookRegistration:
         config_path = tmp_path / "config.toml"
         config_path.write_text('[profiles]\ndefault = "x"\n')
         logger = SimpleNamespace(warning=MagicMock(), info=MagicMock())
-        monkeypatch.setattr(aiusage_mod, "_CODEX_CONFIG", config_path)
-        monkeypatch.setattr(aiusage_mod.log, "bind", lambda **_kwargs: logger)
+        monkeypatch.setattr(codex_backend, "_CODEX_CONFIG", config_path)
+        monkeypatch.setattr(codex_backend.log, "bind", lambda **_kwargs: logger)
         original_write_text = Path.write_text
 
         def fail_write_text(self, *args, **kwargs):
@@ -1251,7 +1273,7 @@ class TestCodexHookRegistration:
 
         monkeypatch.setattr(Path, "write_text", fail_write_text)
 
-        aiusage_mod._register_codex_hook()
+        codex_backend.register_hook()
 
         logger.warning.assert_called_once()
 
@@ -1260,13 +1282,11 @@ class TestHookCli:
     def test_update_config_recovers_from_invalid_existing_json(
         self, tmp_path, monkeypatch
     ):
-        from docking.applets.aiusage import hook
-
         config_path = tmp_path / "dock.json"
         config_path.write_text("{")
-        monkeypatch.setattr(hook, "_config_path", lambda: config_path)
+        monkeypatch.setattr(aiusage_store, "config_path", lambda: config_path)
 
-        hook._update_config(
+        aiusage_store.replace_session(
             session_id="s1",
             model_usage={"gpt-5.4": ModelUsage(input_tokens=12)},
         )
@@ -1277,8 +1297,8 @@ class TestHookCli:
     def test_main_handles_invalid_claude_stdin(self, monkeypatch):
         from docking.applets.aiusage import hook
 
-        monkeypatch.setattr(aiusage_mod.sys, "argv", ["hook", "claude", "Stop"])
-        monkeypatch.setattr(aiusage_mod.sys, "stdin", StringIO("{"))
+        monkeypatch.setattr(hook.sys, "argv", ["hook", "claude", "Stop"])
+        monkeypatch.setattr(hook.sys, "stdin", StringIO("{"))
 
         hook.main()
 
@@ -1286,10 +1306,10 @@ class TestHookCli:
         from docking.applets.aiusage import hook
 
         find_session = MagicMock(return_value=None)
-        monkeypatch.setattr(aiusage_mod.sys, "argv", ["hook"])
-        monkeypatch.setattr(hook.aiusage_state, "find_codex_session", find_session)
+        monkeypatch.setattr(hook.sys, "argv", ["hook"])
+        monkeypatch.setattr(codex_backend, "find_session", find_session)
 
-        hook._handle_codex_turn(json_arg="{")
+        codex_backend.handle_turn(json_arg="{")
 
         find_session.assert_called_once_with(thread_id=None)
